refactor method to helper function
